// examples/typescript/src/_generated/openapi.ts
export interface components {
    schemas: {
// ...
        CapabilityView: {
            /** Name */
            name: string;
            /** Work Unit */
            work_unit: string | null;
            /** Offerings */
            offerings: components["schemas"]["OfferingView"][];
        };
// ...
        OfferingView: {
            /** Id */
            id: string;
            /** Price Per Work Unit Wei */
            price_per_work_unit_wei: string | null;
            /** Work Unit */
            work_unit: string | null;
        };
        /** OrchestratorList */
        OrchestratorList: {
            /** Items */
            items: components["schemas"]["OrchestratorView"][];
        };
        /** OrchestratorView */
        OrchestratorView: {
            /** Eth Address */
            eth_address: string;
            /** Worker Url */
            worker_url: string;
            /** Capabilities */
            capabilities: components["schemas"]["CapabilityView"][];
            /** Signature Status */
            signature_status: string;
            /** Freshness Status */
            freshness_status: string;
        };
// ...
        RouteView: {
            /** Worker Url */
            worker_url: string;
            /** Eth Address */
            eth_address: string;
            /** Capability */
            capability: string;
            /** Offering */
            offering: string;
            /** Price Per Work Unit Wei */
            price_per_work_unit_wei: string;
            /** Work Unit */
            work_unit: string;
            /** Units Per Price */
            units_per_price: number;
            /** Quote Id */
            quote_id: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}